Extract Storage interfaces for `Post` and `User`, moving concrete implementations to `MetaStorage`. Updated modules and tests to reflect changes.
