+++ sources/Post/Module.php
@@ -31,7 +31,7 @@ class Module implements ServiceModule, ExecutableModule
             Post::class => static fn (ContainerInterface $container) => Post::new(
                 $container->get(Repository::class)
             ),
-            Storage::class => static fn () => Storage::new(),
+            Storage::class => static fn () => MetaStorage::new(),
             ItemRegistryKey::class => static fn () => ItemRegistryKey::new(),
             ItemRegistry::class => static fn (
                 ContainerInterface $container
