+++ sources/User/MetaStorage.php
@@ -1,0 +1,45 @@
+<?php
+
+declare(strict_types=1);
+
+namespace SpaghettiDojo\Konomi\User;
+
+/**
+ * @internal
+ */
+class MetaStorage implements Storage
+{
+    public static function new(): MetaStorage
+    {
+        return new self();
+    }
+
+    final private function __construct()
+    {
+    }
+
+    /**
+     * @return array<mixed>
+     */
+    public function read(int $id, string $key): array
+    {
+        if ($id <= 0 || $key === '') {
+            return [];
+        }
+
+        $data = get_user_meta($id, $key, true);
+        return is_array($data) ? $data : [];
+    }
+
+    /**
+     * @param array<mixed> $data
+     */
+    public function write(int $id, string $key, array $data): bool
+    {
+        if ($id <= 0 || $key === '') {
+            return false;
+        }
+
+        return (bool) update_user_meta($id, $key, $data);
+    }
+}
