+++ tests/integration/php/UserRepositoryTest.php
@@ -25,7 +25,7 @@ beforeEach(function (): void {
     $itemRegistryKey = User\ItemRegistryKey::new();
     $this->repository = User\Repository::new(
         User\StorageKey::new('_konomi_items'),
-        User\Storage::new(),
+        User\MetaStorage::new(),
         User\ItemFactory::new(),
         User\ItemRegistry::new($itemRegistryKey),
         User\RawDataAssert::new()
